# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZentProxy API
  version: 0.6.7
  description: Versioned API for ZentProxy proxy management, certificates, routing, analytics, trusted proxies, integrations, user preferences and migration tooling.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ZentProxy API key
    sessionCookie:
      type: apiKey
      in: cookie
      name: zp_session
...
security:
  - bearerAuth: []
paths:
  /user/preferences/language:
    put:
      tags: [System]
      summary: Set the current administrator UI language
      description: Interactive administrator session only. Supported languages are de and en.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [language]
              properties:
                language: { type: string, enum: [de, en] }
      responses:
        '200': { description: Language preference saved }
        '422': { description: Unsupported language }
  /user/preferences/proxy-hosts-view:
    put:
      tags: [System]
      summary: Set the current administrator Proxy Hosts view
      description: Interactive administrator session only.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [view]
              properties:
                view: { type: string, enum: [list, grouped] }
      responses:
        '200': { description: Proxy Hosts view preference saved }
        '422': { description: Unsupported view }
...